+++ HetDepot/Controllers/Tests/KarlijnsTestController.cs
@@ -1,0 +1,28 @@
+namespace HetDepot.Controllers.Tests;
+
+class KarlijnsTestController : Controller
+{
+    /*
+        Karlijn's code from the old Program.cs
+    */
+
+    public override void Execute()
+    {
+        Console.WriteLine("Voer uw code in:");
+        int code = Convert.ToInt32(Console.ReadLine());
+
+        //     Console.WriteLine ("Voer uw code in:");
+        //     var CodeNumber = Console.ReadLine();
+
+        //     int code;
+        //     if (!int.TryParse(CodeNumber, out code))
+
+        //     {
+        //         Console.WriteLine ("Deze code is niet juist, probeer het nog een keer");
+        //         CodeNumber = Console.ReadLine();
+        //     }
+        //     else
+
+        // Console.WriteLine ("Bedankt!");
+    }
+}
