+++ HetDepot/Controllers/Tests/TomsTestController.cs
@@ -1,0 +1,15 @@
+namespace HetDepot.Controllers.Tests;
+
+using HetDepot.Views;
+
+class TomsTestController : Controller
+{
+    /*
+        Tom's code from the old Program.cs
+    */
+
+    public override void Execute()
+    {
+        TourguideView.DisplayTours();
+    }
+}
